refactor: separate hooks and components from UserProfile
